+++ app/src/main/java/com/neome/api/meta/base/dto/DefnLayoutFormWatermark.kt
@@ -17,7 +17,7 @@ interface DefnLayoutFormWatermark
   val bgImageHorizontalPosition: EnumDefnContentAlignment?
   val bgImageVar: FieldDtoImage?
   val bgImageVerticalPosition: EnumDefnContentAlignment?
-  val textOpacityVar: Long?
+  val textOpacityVar: Double?
   val textPatternVar: DefnDtoText?
   val textPositionVar: EnumDefnPlacement?
   val textSizeVar: EnumDefnTextSize?
