+++ app/src/main/java/com/neome/api/meta/base/dto/FieldValueDecimalRange.kt
@@ -7,6 +7,6 @@ package com.neome.api.meta.base.dto
 
 interface FieldValueDecimalRange
 {
-  val maxValue: Long?
-  val minValue: Long?
+  val maxValue: Double?
+  val minValue: Double?
 }
