+++ app/src/main/java/com/neome/api/meta/base/dto/StudioBuildRepeatSettingMonths.kt
@@ -7,9 +7,11 @@ package com.neome.api.meta.base.dto
 
 import com.neome.api.meta.base.Types.AnyTime
 import com.neome.api.meta.base.Types.EnumDefnDateOccurrence
+import com.neome.api.meta.base.dto.StudioBuildRepeatSetting
 
-interface StudioBuildRepeatSettingMonths : StudioBuildRepeatSetting {
-    val customDateSet: List<Long>?
-    val dateOccurrence: EnumDefnDateOccurrence?
-    val setOfTime: List<AnyTime>?
+interface StudioBuildRepeatSettingMonths : StudioBuildRepeatSetting
+{
+  val customDateSet: List<Int>?
+  val dateOccurrence: EnumDefnDateOccurrence?
+  val setOfTime: List<AnyTime>?
 }
