+++ app/src/main/java/com/neome/api/meta/base/dto/StudioBuildRepeatSettingYears.kt
@@ -8,10 +8,12 @@ package com.neome.api.meta.base.dto
 import com.neome.api.meta.base.Types.AnyTime
 import com.neome.api.meta.base.Types.EnumDefnDateOccurrence
 import com.neome.api.meta.base.Types.EnumDefnMonth
+import com.neome.api.meta.base.dto.StudioBuildRepeatSetting
 
-interface StudioBuildRepeatSettingYears : StudioBuildRepeatSetting {
-    val customDateSet: List<Long>?
-    val dateOccurrence: EnumDefnDateOccurrence?
-    val setOfMonth: List<EnumDefnMonth>?
-    val setOfTime: List<AnyTime>?
+interface StudioBuildRepeatSettingYears : StudioBuildRepeatSetting
+{
+  val customDateSet: List<Int>?
+  val dateOccurrence: EnumDefnDateOccurrence?
+  val setOfMonth: List<EnumDefnMonth>?
+  val setOfTime: List<AnyTime>?
 }
