+++ app/src/main/java/com/neome/core/common/serializer/api/meta/base/dto/DefnFieldLogDecimalData.kt
@@ -71,9 +71,9 @@ data class DefnFieldLogDecimalData(
     @Serializable(with = MetaIdFieldSer::class) override val minFieldId: Types.MetaIdField? = null,
     override val minVar: Double? = null,
     override val numberFormat: String? = null,
-    override val numberOfDigitsAfterPeriod: Double? = null,
+    override val numberOfDigitsAfterPeriod: Long? = null,
     @Serializable(with = MetaIdFieldSer::class) override val numberOfDigitsAfterPeriodFieldId: Types.MetaIdField? = null,
-    override val numberOfDigitsAfterPeriodVar: Double? = null,
+    override val numberOfDigitsAfterPeriodVar: Long? = null,
     override val hideInfo: Boolean? = null,
     override val logReadRoleSet: List<@Serializable(with = MetaIdRoleSer::class) Types.MetaIdRole>? = null
 ) : DefnCompSeal, DefnFieldLogDecimal
